Moved ViewModels into separate folder. Created BudgetDropZone component.
