+++ UI/Components/BudgetTransactionDropZone.razor
@@ -1,0 +1,29 @@
+﻿@using UI.Models
+@using UI.Models.ViewModels
+
+<MudDropZone T="TransactionViewModel" Identifier="@Identifier" 
+             Class="rounded mud-background-gray pa-3 ma-3 flex-grow-1" 
+             AllowReorder="false" Style="width: 25%;height: 500px;">
+    <MudText Typo="Typo.h6" Class="mb-4" Style="cssString">
+        @Title @($"{(ActualValue / TotalValue * 100):0.00}%")
+        <MudPaper Class="d-flex justify-center" Elevation="0" Style="position: relative; width: 90%; height: 20px;">
+            <MudProgressLinear Value="@ActualValue" Max="@TotalValue"
+                               Rounded="true" Size="Size.Medium" Class="rounded-progress"
+                               Style="height: 20px;" Color="@TitleColor" />
+            <MudText Style="position: absolute; top: 50%; left: 50%; transform: translate(-50%, -50%); font-weight: bold;">
+                @ActualValue / @TotalValue
+            </MudText>
+        </MudPaper>
+    </MudText>
+</MudDropZone>
+
+@code {
+    [Parameter] public string Identifier { get; set; }
+    [Parameter] public string Title { get; set; }
+    [Parameter] public double ActualValue { get; set; }
+    [Parameter] public double TotalValue { get; set; } 
+    [Parameter] public string Color { get; set; } 
+    [Parameter] public MudBlazor.Color TitleColor { get; set; }
+
+    private string cssString => $"text-align:center;color:{Color}";
+}
