+++ UI/Components/Dialogs/EditMonthPatternDialog.razor
@@ -1,4 +1,6 @@
 ﻿@using UI.Models
+@using UI.Models.ViewModels
+
 <MudDialog>
     <DialogContent>
         <MudText @bind-vale="contextModel.Pattern.Name" />
