+++ UI/Components/Dialogs/PatternDialog.razor
@@ -1,4 +1,5 @@
 ﻿@using UI.Models
+@using UI.Models.ViewModels
 <MudDialog>
     <DialogContent>
         <MudStack Row="true" AlignItems="AlignItems.Center" Justify="Justify.Center">
