+++ UI/Models/PatternValuesModel.cs
@@ -1,0 +1,13 @@
+﻿namespace UI.Models
+{
+    public class PatternValuesModel
+    {
+        public double ActualValueSaves { get; set; }
+        public double ActualValueFees { get; set; }
+        public double ActualValueEntertainment { get; set; }
+        public double TotalValueSaves { get; set; }
+        public double TotalValueFees { get; set; }
+        public double TotalValueEntertainment { get; set; }
+
+    }
+}
