+++ UI/Pages/MyPages/OptionsPage/Income.razor
@@ -1,5 +1,6 @@
 ﻿@page "/options/income"
-@using UI.Models
+@using UI.Models;
+@using UI.Models.ViewModels
 
 
 <UI.Components.OptionNavigationComponent />
