+++ UI/Pages/MyPages/OptionsPage/Pattern.razor
@@ -1,5 +1,6 @@
 ﻿@page "/options/pattern"
-@using UI.Models
+@using UI.Models;
+@using UI.Models.ViewModels
 
 <UI.Components.OptionNavigationComponent />
 <MudPaper Class="ma-1 pa-2" Elevation="5" style="text-align:center;">
